+++ src/Modal/Footer.js
@@ -15,6 +15,7 @@ ModalFooter.propTypes = {
   extraClass: PT.string,
   /** Childen nodes */
   children: PT.oneOfType([
+    PT.string,
     PT.arrayOf(PT.node),
     PT.node
   ]).isRequired