+++ src/Modal/Header.js
@@ -2,46 +2,38 @@ import React from 'react'
 import {default as PT} from 'prop-types'
 import cx from 'classnames'
 
-const ModalTitle = ({ title, htmlTitle=false }) => {
-  return htmlTitle ?
-    <h2 className="ola_modal-title" dangerouslySetInnerHTML={{__html: title}} /> :
-    <h2 className="ola_modal-title">{title}</h2>
-}
-
-const ModalIntro = ({ intro, htmlIntro=false }) => {
-  return htmlIntro ?
-    <p className="ola_modal-intro" dangerouslySetInnerHTML={{__html: intro}} /> :
-    <p className="ola_modal-intro">{intro}</p>
-}
-
-const ModalHeader = ({ title, htmlTitle, intro, htmlIntro, children, ...props }) => {
+const ModalHeader = ({ title, intro, children, extraClass, ...props }) => {
   return (
-    <header className={cx('ola_modal-header', children && 'has-extra')} {...props}>
-      <ModalTitle title={title} htmlTitle={htmlTitle} />
-      { intro && <ModalIntro intro={intro} htmlIntro={htmlIntro} /> }
+    <header className={cx('ola_modal-header', children && 'has-extra', extraClass)} {...props}>
+      <h2 className="ola_modal-title">{title}</h2>
+      { intro && <p className="ola_modal-intro">{intro}</p> }
       { children && <div className="ola_modal-extra">{ children }</div> }
     </header>
   )
 }
 
 ModalHeader.defaultProps = {
-  intro: null,
-  htmlTitle: false,
-  htmlIntro: false,
-  children: undefined
+  intro: null
 }
 
 ModalHeader.propTypes = {
+  /** Extra className */
+  extraClass: PT.string,
   /** Title of header */
-  title: PT.string.isRequired,
-  /** Title support HTML tags */
-  htmlTitle: PT.bool,
-  /** Intro support HTML tags */
-  htmlIntro: PT.bool,
+  title: PT.oneOfType([
+    PT.string,
+    PT.arrayOf(PT.node),
+    PT.node
+  ]).isRequired,
   /** Intro text of header */
-  intro: PT.string,
+  intro: PT.oneOfType([
+    PT.string,
+    PT.arrayOf(PT.node),
+    PT.node
+  ]),
   /** Childen nodes */
   children: PT.oneOfType([
+    PT.string,
     PT.arrayOf(PT.node),
     PT.node
   ])